have started to create main content of frondend

+++ src/controllers/todo.controller.js
@@ -3,6 +3,8 @@ import { ToDo } from "../schema/todo.schema.js";
 export const createTodo = async (req, res) => {
   let { title, content, table, priority, deadLine, status } = req.body;
 
+  const userId = req.user._id;
+
   let dateNow = new Date();
   const dateWithNewYear = dateNow.setFullYear(dateNow.getFullYear() + 1);
 
@@ -15,47 +17,57 @@ export const createTodo = async (req, res) => {
     priority,
     deadLine,
     status,
+    userId,
   });
   res.status(201).json(toDo);
 };
 
 export const getTodos = async (req, res) => {
-  const toDos = await ToDo.find();
+  const userId = req.user._id;
+  const toDos = await ToDo.find({ userId });
 
   return res.status(200).send(toDos);
 };
 
 export const getTodoByID = async (req, res) => {
+  const userId = req.user._id;
   const { todoId } = req.params;
 
-  const toDo = await ToDo.findOne({ _id: todoId });
+  const toDo = await ToDo.findOne({ _id: todoId, userId });
 
   return res.status(200).send(toDo);
 };
 
 export const deleteTodoByID = async (req, res) => {
+  const userId = req.user._id;
   const { todoId } = req.params;
 
-  const toDo = await ToDo.findByIdAndDelete({ _id: todoId });
+  const toDo = await ToDo.findOneAndDelete({ _id: todoId, userId });
 
   return res.status(200).send(toDo);
 };
 
 export const updateTodoByID = async (req, res) => {
+  const userId = req.user._id;
   const { todoId } = req.params;
 
-  const updatedTodo = await ToDo.findByIdAndUpdate({ _id: todoId }, req.body, {
-    new: true,
-  });
+  const updatedTodo = await ToDo.findOneAndUpdate(
+    { _id: todoId, userId },
+    req.body,
+    {
+      new: true,
+    }
+  );
 
   return res.status(200).send(updatedTodo);
 };
 
 export const getTodosByTable = async (req, res) => {
   try {
+    const userId = req.user._id;
     const { table } = req.params;
 
-    const todosByTable = await ToDo.find({ table });
+    const todosByTable = await ToDo.find({ table, userId });
 
     res.send(todosByTable);
   } catch (error) {
@@ -65,9 +77,10 @@ export const getTodosByTable = async (req, res) => {
 
 export const deleteTododsByTable = async (req, res) => {
   try {
+    const userId = req.user._id;
     const { table } = req.params;
 
-    await ToDo.deleteMany({ table });
+    await ToDo.deleteMany({ table, userId });
 
     res.status(200).send(`todos by table ${table} were delete`);
   } catch (error) {
@@ -77,10 +90,11 @@ export const deleteTododsByTable = async (req, res) => {
 
 export const updateTodosTableName = async (req, res) => {
   try {
+    const userId = req.user._id;
     const { table } = req.params;
 
     const todosByTable = await ToDo.updateMany(
-      { table },
+      { table, userId },
       { table: req.body.table }
     );
